Remove space from text string
The text string shouldn’t have a space in it. It was added to separate the icon from the text. I’ve added the space after the <span> instead and removed it from the text string. Going forward it will look exactly the same as it did before (2 spaces will show as one) so doesn’t cause any backwards compatibility concerns.